+++ index.html
@@ -23,6 +23,6 @@
     <!-- <div class="add-form promise-add hidden">Добавляю комментарий...</div> -->
   </div>
   
-  <script src="script.js"></script>
+  <script type="module" src="script.js"></script>
 </body>
 </html>
