+++ script.js
@@ -1,23 +1,21 @@
+import { getTodos, postTodos } from "./api.js";
+import { formatDate } from "./data.js";
+import { renderComments } from "./renderComment.js"
+
 "use strict";
 console.log("It works!");
 const buttonElement = document.getElementById("add-button");
-const commentElement = document.getElementById("list");
+// const commentElement = document.getElementById("list");
 const nameInputElement = document.getElementById("name-input");
 const textInputElement = document.getElementById("text-input");
 const container = document.querySelector(".container");
 const addForm = document.querySelector(".add-form");
 
+
 // Подключаем приложение комментариев к API
 // fetch - запускает выполнение запроса к api
 const fetchPromise = () => {
-  return fetch('https://wedev-api.sky.pro/api/v1/nadya-terleeva/comments', {
-    method: "GET"
-  })
-    // подписываемся на успешное завершение запроса с помощью then
-    .then((response) => {
-
-      const jsonPromise = response.json();
-      jsonPromise.then((responseData) => {
+ getTodos().then((responseData) => {
         // Преобразовываем данные из формата API в формат приложения
         const appComments = responseData.comments.map((comment) => {
           return {
@@ -34,35 +32,16 @@ const fetchPromise = () => {
         })
 
         comments = appComments;
-        renderComments();
+        renderComments({ comments });
       });
-    });
-}
-function formatDate(myDate) {
-  let date = myDate.getDate();
-  let month = myDate.getMonth() + 1;
-  let hour = myDate.getHours();
-  let minute = myDate.getMinutes();
+    };
 
-  if (date < 10) {
-    date = "0" + date;
-  }
-  if (month < 10) {
-    month = "0" + month;
-  }
-  if (hour < 10) {
-    hour = "0" + hour;
-  }
-  if (minute < 10) {
-    minute = "0" + minute;
-  }
-  return `${date}.${month}.${myDate.getFullYear().toString().substr(-2)} ${hour}:${minute}`;
-};
+    let currentDate = new Date(); // Создается объект текущей даты
+    formatDate(currentDate); // Передается аргумент с текущей датой
 
-let comments = [
-]
+let comments = [];
 
-function appComment(userName, userComment, userData) {
+function appComment() {
   addForm.classList.add("hidden");
 
   let elem = document.createElement("p"); // Добавляем созданный элемент
@@ -70,27 +49,21 @@ function appComment(userName, userComment, userData) {
   elem.classList.add("commentElem");
   container.appendChild(elem);
 
-  const url = 'https://wedev-api.sky.pro/api/v1/nadya-terleeva/comments';
-  fetch(url, {
-    method: 'POST',
-    body: JSON.stringify({
-      name: userName,
-      text: userComment,
-      data: userData,
-      forceError: true
-    }),
-  })
-  .then((response) => {
-    if (response.status === 400) {
-      alert("Имя или комментарий должены содержать хотя бы 3 символа");
-      throw new Error("Имя или комментарий должены содержать хотя бы 3 символа");
-    }
-    if (response.status === 500) {
+    postTodos({
+      userName: nameInputElement.value,
+      userComment: textInputElement.value,
+      userData: formatDate(new Date()),
+    })
+    .then((response) => {
+      if (response.status === 400) {
+        throw new Error("Плохой запрос");
+      }
+      if (response.status === 500) {
         throw new Error("Сервер сломался");
-    }
-    return response.json();
-    })  
-    .then(() => {
+      }
+      return response.json();
+    })      
+     .then(() => {
       fetchPromise();
     })
     .then(() => {
@@ -110,21 +83,26 @@ function appComment(userName, userComment, userData) {
 
       // Если сервер сломался, то просим попробовать позже
       if (error.message === "Сервер сломался") {
-      alert("Сервер сломался, попробуй позже");
-      return;
-    } 
-    if (error.message === "Failed to fetch") {
-      alert("Кажется, у вас сломался интернет, попробуйте позже");
-      return;
-    }
-        // Во всех остальных случаях просто вывдим ошибку
-    console.warn(error);
+        alert("Сервер сломался, попробуй позже");
+        return;
+      }
+      if (error.message === "Failed to fetch") {
+        alert("Кажется, у вас сломался интернет, попробуйте позже");
+        return;
+      }
+      if (error.message === "Плохой запрос") {
+        alert("Имя или комментарий должены содержать хотя бы 3 символа");
+        return;
+      }
+      // Во всех остальных случаях просто вывдим ошибку
+      console.warn(error);
     });
-  renderComments();
+
+  renderComments({ comments });
 }
 
 buttonElement.addEventListener("click", () => {
-  
+
   nameInputElement.classList.remove("error")
   if (nameInputElement.value === "") {
     nameInputElement.classList.add("error");
@@ -144,72 +122,5 @@ buttonElement.addEventListener("click", () => {
   appComment(userName, userComment, userData);
 });
 
-const delay = (ms) => {
-  return new Promise((resolve) => {
-    setTimeout(resolve, ms);
-  });
-};
-
-const initLikeButtonListeners = () => {
-  for (const likeButton of document.querySelectorAll(".like-button")) {
-    likeButton.addEventListener("click", (event) => {
-      console.log(likeButton);
-      event.stopPropagation();
-      const index = likeButton.dataset.index;
-      const comment = comments[index];
-
-      delay(2000).then(() => {
-       if (comment.likes) {
-        comment.likes = false;
-        comment.numberLikes--;
-     } else { 
-      comment.likes = true;
-      comment.numberLikes++;
-      };
-        renderComments();
-    })
-    });
-  };
-};
-
-const renderComments = () => {
-  if (comments.length === 0) {
-    commentElement.textContent = "Пожалуйста подождите, комментарии загружаются...";
-    return;
-  }
-
-  const commentsHtml = comments.map((comment, index) => {
-    const commentTextQuotes = comment.comment.replaceAll("QUOTE_BEGIN", "<div class='quote'>").replaceAll("QUOTE_END", "</div>");
-    const commentNameSafe = comment.name.replace(/</g, '&lt;').replace(/>/g, '&gt;');
-    return `
-    <li class="comment" id="comment-add" data-index=${index}>
-      <div class="comment-header">
-        <div>${commentNameSafe}</div>
-        <div>${comment.data}</div>
-      </div>
-      <div class="comment-body">
-        <div class="comment-text" >${commentTextQuotes}</div>
-      </div>
-      <div class="comment-footer">
-        <div class="likes">
-          <span class="likes-counter" data-index=${index}>${comment.numberLikes}</span>
-          <button class="like-button${comment.likes ? " -active-like" : ""}" data-index=${index}></button>
-        </div>
-      </div>
-    </li>`
-  }).join("");
-
-  commentElement.innerHTML = commentsHtml;
-  initLikeButtonListeners();
-  const commentElements = document.querySelectorAll('.comment');
-  for (const commentElement of commentElements) {
-    const index = commentElement.dataset.index;
-    commentElement.addEventListener('click', () => {
-      const { name, comment } = comments[index];
-      textInputElement.value = 'QUOTE_BEGIN' + ' (' + name + ') ...' + comment + '... ' + 'QUOTE_END' + ' ';
-    });
-  };
-}
-
-renderComments();
+renderComments({ comments });
 fetchPromise();
